[rename] TTS 리스트, 테이블 컴포넌트 tts 폴더로 이사 #43
